Use int values as click acc

+++ application/game/upgrade_points_calculator.py
@@ -6,8 +6,8 @@ from models import UpgradePurchase
 
 class UpgradePointsCalculator:
     @staticmethod
-    def calculate_points(upgrade_purchases: List[UpgradePurchase]) -> float:
-        return float(
+    def calculate_points(upgrade_purchases: List[UpgradePurchase]) -> int:
+        return int(
             math.prod([
                 upgrade_purchase.upgrade.multiplier
                 for upgrade_purchase in upgrade_purchases
